requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
anyio==4.3.0
apache-airflow==2.8.3
apache-airflow-providers-common-io==1.3.0
apache-airflow-providers-common-sql==1.11.1
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-google==10.16.0
apache-airflow-providers-http==4.10.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-smtp==1.6.1
apache-airflow-providers-sqlite==3.7.1
apispec==6.5.0
argcomplete==3.2.3
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
blinker==1.7.0
cachelib==0.9.0
cachetools==5.3.3
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.2
cryptography==41.0.7
db-dtypes==1.2.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.2.0
Flask==2.2.5
Flask-AppBuilder==4.3.11
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.2.0
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.2.0
gcsfs==2024.2.0
google-ads==23.1.0
google-analytics-admin==0.22.7
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.43.0
google-cloud-appengine-logging==1.4.3
google-cloud-audit-log==0.2.5
google-cloud-automl==2.13.3
google-cloud-batch==0.17.14
google-cloud-bigquery==3.19.0
google-cloud-bigquery-datatransfer==3.15.1
google-cloud-bigtable==2.23.0
google-cloud-build==3.23.3
google-cloud-compute==1.18.0
google-cloud-container==2.43.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.18.3
google-cloud-dataflow-client==0.8.10
google-cloud-dataform==0.5.9
google-cloud-dataplex==1.12.3
google-cloud-dataproc==5.9.3
google-cloud-dataproc-metastore==1.15.3
google-cloud-dlp==3.16.0
google-cloud-kms==2.21.3
google-cloud-language==2.13.3
google-cloud-logging==3.9.0
google-cloud-memcache==1.9.3
google-cloud-monitoring==2.19.3
google-cloud-orchestration-airflow==1.12.1
google-cloud-os-login==2.14.3
google-cloud-pubsub==2.20.1
google-cloud-redis==2.15.3
google-cloud-resource-manager==1.12.3
google-cloud-run==0.10.5
google-cloud-secret-manager==2.18.3
google-cloud-spanner==3.43.0
google-cloud-speech==2.25.1
google-cloud-storage==2.15.0
google-cloud-storage-transfer==1.11.3
google-cloud-tasks==2.16.3
google-cloud-texttospeech==2.16.3
google-cloud-translate==3.15.3
google-cloud-videointelligence==2.13.3
google-cloud-vision==3.7.2
google-cloud-workflows==1.14.3
google-crc32c==1.5.0
google-re2==1.1
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpc-interceptor==0.15.4
grpcio==1.62.1
grpcio-gcp==0.2.2
grpcio-status==1.62.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.3
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
json-merge-patch==0.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
limits==3.10.0
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==24.2.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
numpy==1.24.4
oauthlib==3.2.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
ordered-set==4.1.0
packaging==23.2
pandas==2.1.4
pandas-gbq==0.22.0
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.4.0
prison==0.2.1
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydata-google-auth==1.8.2
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.15.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.4.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.1
rich-argparse==1.4.0
rpds-py==0.18.0
rsa==4.9
setproctitle==1.3.3
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
sqlalchemy-bigquery==1.10.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.6.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.0
typing_extensions==4.10.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal-pathlib==0.1.4
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.17.0
